Refactored Flask routes to Pyramid
