+++ src/flask_backup/initialize_db.py
@@ -1,0 +1,5 @@
+from models import init_db
+
+# Initialize the database
+if __name__ == '__main__':
+    init_db()
