+++ src/flask_backup/requirements.txt
@@ -1,0 +1,4 @@
+setuptools==44.1.1
+Werkzeug==0.6.1
+Jinja2==2.4
+Flask==0.5
