+++ src/flask_backup/static/style.css
@@ -1,0 +1,81 @@
+/* Website Defaults */
+body {
+    font-family: 'Trebuchet MS', sans-serif;
+    font-size: 14px;
+}
+
+div.page div.body {
+    padding: 10px;
+}
+
+div.page div.footer {
+    background: #eee;
+    color: #888;
+    padding: 5px 10px;
+    font-size: 12px;
+}
+
+
+ul.flashes {
+    list-style: none;
+    margin: 10px 10px 0 10px;
+    padding: 0;
+}
+
+ul.flashes li {
+    background: #B9F3ED;
+    border: 1px solid #81CEC6;
+    -moz-border-radius: 2px;
+    -webkit-border-radius: 2px;
+    padding: 4px;
+    font-size: 13px;
+}
+
+div.error {
+    margin: 10px 0;
+    background: #FAE4E4;
+    border: 1px solid #DD6F6F;
+    -moz-border-radius: 2px;
+    -webkit-border-radius: 2px;
+    padding: 4px;
+    font-size: 13px;
+}
+
+
+/* Custom Styles */
+nav {
+    white-space: nowrap;
+    text-align: center;
+}
+
+nav h1, nav a {
+    display: inline-block;
+    margin-right: 20px;
+}
+
+#nav-login, #nav-register, #nav-logout {
+    float: right;
+}
+
+input {
+    width: 85%;
+    padding: 10px;
+    margin: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+}
+
+button {
+    width: 10%;
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    background-color: #f0f0f0;
+    cursor: pointer;
+    margin-left: 10px;
+}
+
+button:active {
+    background-color: #e0e0e0;
+    box-shadow: inset 0 3px 5px rgba(0, 0, 0, 0.125);
+}
