+++ src/flask_backup/templates/register.html
