+++ src/schema.sql
@@ -1,7 +1,7 @@
-DROP TABLE IF EXISTS users;
+ DROP TABLE IF EXISTS users;
 
-CREATE TABLE IF NOT EXISTS users (
-  id INTEGER PRIMARY KEY AUTOINCREMENT,
+CREATE TABLE users (
+  id INTEGER PRIMARY KEY,
   username TEXT NOT NULL UNIQUE,
   email TEXT NOT NULL UNIQUE,
   password TEXT NOT NULL
@@ -12,7 +12,7 @@ INSERT INTO users (username, email, password)
     VALUES ('admin', 'keamonk1@stud.kea.dk', '5f4dcc3b5aa765d61d8327deb882cf99');
 
 
-CREATE TABLE IF NOT EXISTS pages (
+CREATE TABLE pages (
     title TEXT PRIMARY KEY UNIQUE,
     url TEXT NOT NULL UNIQUE,
     language TEXT NOT NULL CHECK(language IN ('en', 'da')) DEFAULT 'en', -- How you define ENUM type in SQLite
